requirements.txt
discord.py
openai
requests
pillow